cloud_webserver/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
blinker==1.7.0
boto3==1.34.71
botocore==1.34.71
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.7
lz4==4.3.3
MarkupSafe==2.1.5
mcap==1.1.1
mcap-protobuf-support==0.4.1
numpy==1.26.4
prompt-toolkit==3.0.43
protobuf==5.26.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
s3transfer==0.10.1
scipy==1.12.0
six==1.16.0
typing==3.7.4.3
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
zstandard==0.22.0
